requirements.txt
Django>=3.2.0,<4.0.0
python-dotenv
djangorestframework==3.13.1
django-filter==2.4.0
pillow
djangorestframework-simplejwt==4.8.0
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.2
djoser==2.1.0